feat: add an entrypoint which embeds the wasm instead of requiring it to be fetched over the network when deployed as a Netlify Edge Function
Netlify Edge Functions are not able to read the file-system, which means we can't read the generated `.wasm` file during execution, we currently fetch the wasm file over the network but this request could fail or have an incorrect response, leading to a runtime error when attempting to instantiate the WebAssembly Module. To work-around this, we when building the `.wasm` file we now also build a `.ts` file which includes the exact same bytes as the wasm file, but within a JavaScipt UInt8Array, which we can then instantiate as a WebAssembly Module.